+++ src/WordPress/Http/Route.php
@@ -81,8 +81,6 @@ class Route {
 		return true;
 	}
 
-	
-
 	function when($callback)
 	{
 		$this->permissionCallback = $callback;