+++ src/WordPress/TestCaseResponse.php
@@ -1,0 +1,10 @@
+<?php
+namespace FatPanda\WordPress;
+
+/**
+ * Wrap this class around some response content, and then use its
+ * methods to make assertions about the contents the response contains.
+ */
+class TestCaseResponse extends PHPUnit_Framework_Assert {
+
+}