+++ tests/test-bootstrap.php
@@ -6,8 +6,6 @@ use FatPanda\Illuminate\WordPress\TestCase;
  */
 class TestBootstrap extends TestCase {
 
-	protected $plugin = 'test-plugin';
-
 	/**
 	 * Make sure that our test framework is in place
 	 */