+++ tests/test-core.php
@@ -6,17 +6,21 @@ use FatPanda\Illuminate\WordPress\TestCase;
  */
 class TestCore extends TestCase {
 
-	protected $plugin = 'test-plugin';
+	function setUp()
+	{
+		parent::setUp();
+		$this->plugin = plugin('test-plugin');
+	}
 
 	/**
 	 * Make sure our access to plugin metadata is working
 	 */
 	function testPluginData()
 	{	
-		$data = $this->plugin()->getPluginData();
+		$data = $this->plugin->getPluginData();
 		$this->assertNotEmpty( $data );
 		$this->assertEquals( $data['Name'], 'Test Plugin' );
-		$this->assertEquals( $this->plugin()->getPluginData('Name'), 'Test Plugin' );
+		$this->assertEquals( $this->plugin->getPluginData('Name'), 'Test Plugin' );
 	}
 
 }