+++ tests/test-plugin/src/Models/PostWithAlternativeConnection.php
@@ -1,0 +1,10 @@
+<?php
+namespace FatPanda\WordPress\Models;
+
+use FatPanda\Illuminate\WordPress\Models\Post;
+
+class PostWithAlternativeConnection extends Post {
+
+	protected $connection = 'alternative';
+
+}