+++ tests/test-plugin/src/routes.php
@@ -1,1 +1,4 @@
 <?php
+$router->get('/plugin-data/{property?}', function($request) use ($plugin) {
+	return $plugin->getPluginData($request['property']);
+});